[change] int instead size_t

+++ 0x11-singly_linked_lists/1-list_len.c
@@ -6,7 +6,7 @@
  */
 size_t list_len(const list_t *h)
 {
-	size_t length = 0;
+	int length = 0;
 
 	if (h)
 	{
